Favorite redirect for guests
Currently favoriting stores and projects is not working unless you are logged
in. Previously there was partial support for this, but it required a special
create account component and if you already had an account there was no
corresponding login component.

Since support for a redir parameter has been added to the Add to Favorites CMS
Generator, update favorite-button.js so that if the standard login and create
account components redirect to that the favorite will be set and the user will
be returned to the page they were viewing.

GIV-146 Shouldn't the ability to favorite a store be available for non
logged-in users?